Close the SEV/TRP gate-coverage gap by classification
Every SEV and TRP rule now names a gate or carries a labeled exclusion.
Three reuse mappings extend existing gates (SEV-027 into SEV-151,
SEV-031's range made explicit in SEV-124, SEV-090 into SEV-133 and
SEV-150a/b/d); SEV-157/158 are the Complex reject-before-backend and
term-reuse NotImplemented-until-exists gates; SEV-159..162 and
TRP-100..103 are governance-labeled static deletion and authority scans;
SEV-163..168 and TRP-104/106 are the executable gates, with SEV-168
requiring every identity family swapped/duplicated/omitted; TRP-105 is
the one exhaustive ProductKind x RootKind matrix closing TRP-019/022/
049/050 together. TRP-004/040/041 are labeled governance/reversal
exclusions rather than invented runtime tests. The coverage checker now
expands ID ranges and counts only rows claiming implementation.
Uncovered: SEV 0, TRP 0, SDO 0.
