+++ crates/rumoca-sim/src/native_execution.rs
@@ -18,7 +18,7 @@ impl rumoca_solver::CompiledSolveExpression for CraneliftExpression {
         y: &[f64],
         p: &[f64],
         t: f64,
-        external_tables: &[rumoca_core::ExternalTableData],
+        external_tables: &[rumoca_ir_solve::ExternalTableData],
         out: &mut [f64],
     ) -> Result<(), String> {
         self.0
@@ -34,7 +34,7 @@ impl rumoca_solver::CompiledSolveJacobianExpression for CraneliftJacobianExpress
         p: &[f64],
         t: f64,
         seed: &[f64],
-        external_tables: &[rumoca_core::ExternalTableData],
+        external_tables: &[rumoca_ir_solve::ExternalTableData],
         out: &mut [f64],
     ) -> Result<(), String> {
         self.0
@@ -49,7 +49,7 @@ impl rumoca_solver::CompiledSolveAssignmentSchedule for CraneliftAssignmentSched
         y: &mut [f64],
         p: &[f64],
         t: f64,
-        external_tables: &[rumoca_core::ExternalTableData],
+        external_tables: &[rumoca_ir_solve::ExternalTableData],
     ) -> Result<(), String> {
         self.0
             .call_with_external_tables(y, p, t, external_tables)
